bundle FE code for standalone distribution

+++ client/build.sh
@@ -6,35 +6,47 @@ set -eo pipefail
 # `.wasm` files to the `public` directory so that they're available to the
 # browser at runtime.
 
-echo "⚡️ Initializing..."
-mkdir -p public/editor/themes
-mkdir -p public/editor/themes/custom
-mkdir -p public/editor/themes/monaco-themes
+cd $(dirname "$0")
 
-echo "⚡️ Copying custom themes..."
-cp src/components/editor/themes/* public/editor/themes/custom
+if [ ! -d "public/editor/themes" ]; then
+  echo "⚡️ Initializing..."
+  mkdir -p public/editor/themes
+fi
 
-echo "⚡️ Caching monaco-themes..."
-curl -Lo public/editor/themes/monaco-themes.json https://unpkg.com/monaco-themes/themes/themelist.json
-jq -r 'to_entries[] | "\(.key) \(.value)"' public/editor/themes/monaco-themes.json | while read -r theme label; do
-  echo "Fetching \"$label\" to themes/monaco-themes/$theme.json..."
-  curl --progress-bar -Lo public/editor/themes/monaco-themes/$theme.json "https://unpkg.com/monaco-themes/themes/${label// /%20}.json"
-done
+if [ ! -d "public/editor/themes/custom" ]; then
+  echo "⚡️ Copying custom themes..."
+  mkdir -p public/editor/themes/custom
+  cp src/components/editor/themes/* public/editor/themes/custom
+fi
 
-echo "⚡️ Installing web-tree-sitter..."
-deno install
-cp node_modules/web-tree-sitter/tree-sitter.wasm public/editor/tree-sitter.wasm
-cp node_modules/web-tree-sitter/tree-sitter.wasm.map public/editor/.tree-sitter.wasm.map
+if [ ! -d "public/editor/themes/monaco-themes" ]; then
+  echo "⚡️ Caching monaco-themes..."
+  mkdir -p public/editor/themes/monaco-themes
+  curl -Lo public/editor/themes/monaco-themes.json https://unpkg.com/monaco-themes/themes/themelist.json
+  jq -r 'to_entries[] | "\(.key) \(.value)"' public/editor/themes/monaco-themes.json | while read -r theme label; do
+    echo "Fetching \"$label\" to themes/monaco-themes/$theme.json..."
+    curl --progress-bar -Lo public/editor/themes/monaco-themes/$theme.json "https://unpkg.com/monaco-themes/themes/${label// /%20}.json"
+  done
+fi
 
-echo "⚡️ Installing tree-sitter-sql..."
-git clone https://github.com/DerekStride/tree-sitter-sql
-cd tree-sitter-sql
-npx tree-sitter-cli generate
-npx tree-sitter-cli build --wasm
-cp tree-sitter-sql.wasm ../public/editor/tree-sitter-sql.wasm
-cd ..
+if [ ! -f "public/editor/tree-sitter.wasm" ]; then
+  echo "⚡️ Installing web-tree-sitter..."
+  deno install
+  cp node_modules/web-tree-sitter/tree-sitter.wasm public/editor/tree-sitter.wasm
+  cp node_modules/web-tree-sitter/tree-sitter.wasm.map public/editor/.tree-sitter.wasm.map
+fi
 
-echo "⚡️ Cleaning up..."
-rm -rf tree-sitter-sql
+if [ ! -f "public/editor/tree-sitter-sql.wasm" ]; then
+  echo "⚡️ Installing tree-sitter-sql..."
+  git clone https://github.com/DerekStride/tree-sitter-sql
+  cd tree-sitter-sql
+  npx tree-sitter-cli generate
+  npx tree-sitter-cli build --wasm
+  cp tree-sitter-sql.wasm ../public/editor/tree-sitter-sql.wasm
+  cd ..
+
+  echo "⚡️ Cleaning up..."
+  rm -rf tree-sitter-sql
+fi
 
 echo "✅ Done!"
